Move lazy_init under scipp folder again
