+++ src/binding/python/openpmd_api/scipp/lazy_init.py
@@ -1,8 +1,7 @@
 def series_to_scipp(series):
-
     import scipp
 
-    from .scipp import DataLoader
+    from . import DataLoader
 
     dl = DataLoader(series)
     return dl
